app/requirements.txt
alembic==1.0.7
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-translate==1.3.3
googleapis-common-protos==1.5.8
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
PyJWT==1.7.1
PySocks==1.6.8
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.2.18
twilio==6.24.1
urllib3==1.24.1
Werkzeug==0.14.1
WTForms==2.2.1
